refacto: move compile_bytes in Matcher
